requirements.txt
alembic==1.9.2
amqp==5.1.1
bcrypt==4.0.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
debtcollector==2.5.0
decorator==5.1.1
dnspython==2.3.0
eventlet==0.33.3
extras==1.0.0
fasteners==0.18
fixtures==4.0.1
futurist==2.4.1
greenlet==2.0.1
idna==3.4
iso8601==1.1.0
Jinja2==3.1.2
kombu==5.2.4
Mako==1.2.4
MarkupSafe==2.1.2
monotonic==1.6
msgpack==1.0.4
msgpack-python==0.5.6
netaddr==0.8.0
netifaces==0.11.0
oslo.concurrency==5.0.1
oslo.config==8.5.1
oslo.context==5.0.0
oslo.db==8.5.1
oslo.i18n==5.1.0
oslo.log==4.4.0
oslo.messaging==14.1.0
oslo.metrics==0.5.0
oslo.middleware==5.0.0
oslo.serialization==5.0.0
oslo.service==2.5.0
oslo.utils==4.8.0
oslo.versionedobjects==2.4.0
packaging==23.0
Paste==3.5.2
PasteDeploy==3.0.1
pbr==5.11.1
prometheus-client==0.15.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
repoze.lru==0.7
requests==2.28.2
rfc3986==2.0.0
Routes==2.5.1
six==1.16.0
SQLAlchemy==1.4.46
sqlalchemy-filters==0.12.0
sqlalchemy-migrate==0.13.0
sqlparse==0.4.3
statsd==4.0.1
stevedore==4.1.1
Tempita==0.5.2
testresources==2.0.1
testscenarios==0.5.0
testtools==2.5.0
urllib3==1.26.14
vine==5.0.0
WebOb==1.8.7
wrapt==1.14.1
yappi==1.4.0
